requirements.txt
pip
wheel
setuptools
tflite_runtime-2.6.0-cp39-none-linux_aarch64.whl
colorama==0.4.4
librosa
pytz
tzlocal
pandas
seaborn
streamlit==1.19.0
plotly
apprise==1.2.1
paho-mqtt
pytest==7.1.2
pytest-mock==3.7.0
suntime
pyserial
altair<5